# cloudify/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Cloudify API
  description: |
    Cloudify is a self-hosted deployment platform for modern web applications.
    This API provides programmatic access to manage projects, deployments, domains,
    environment variables, serverless functions, and more.

    ## Authentication
    All API requests require authentication using a Bearer token.
    Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-api-token>
    ```

    ## Rate Limiting
    API requests are rate limited to 1000 requests per minute per API key.

    ## Errors
    The API uses standard HTTP response codes to indicate success or failure.
    - 2xx: Success
    - 4xx: Client error (bad request, unauthorized, not found)
    - 5xx: Server error
  version: 1.0.0
  contact:
    name: Cloudify Support
    email: support@cloudify.dev
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://cloudify.tranthachnguyen.com/api
    description: Production server
  - url: http://localhost:3000/api
    description: Development server
...
paths:
...
  /webhooks/github:
    post:
      tags: [Webhooks]
      summary: GitHub webhook endpoint
      operationId: githubWebhook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Webhook processed
  /webhooks/gitlab:
    post:
      tags: [Webhooks]
      summary: GitLab webhook endpoint
      operationId: gitlabWebhook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Webhook processed
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  - bearerAuth: []